requirements.txt
asgiref==3.5.0
bs4==0.0.1
cloudinary==1.29.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.50.0
django-currencies==0.10.1
jsonfield==2.1.1
oauthlib==3.2.0
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
